pyproject.toml
[project]
name = "datalabeling"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fadel Seydou", email = "fadel.seydou@gmail.com" }
]
requires-python = "~=3.11"
dependencies = [
    "datargs>=1.1.0",
    "dotenv>=0.9.9",
    "label-studio-ml>=1.0.9",
    "label-studio-sdk>=1.0.10",
    "label-studio-converter",
    "lightning>=2.5.0.post0",
    "matplotlib>=3.10.1",
    "mlflow==2.13.2",
    "notebook>=7.3.2",
    "numpy>=1.26.4",
    "openpyxl>=3.1.5",
    "pillow>=11.1.0",
    "sahi>=0.11.22",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "wandb==0.18.7",
    "black>=25.1.0",
    "isort>=5.13.2",
    "ruff>=0.9.10",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "nbqa>=1.9.1",
    "geopy>=2.4.1",
    "utm>=0.8.1",
    "litserve>=0.2.8",
    "ray[data,serve,train,tune]>=2.44.1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "datalabeling",
    "ruff>=0.9.10",
]

[tool.uv.sources]
datalabeling = { workspace = true }

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = ["build", "dist", ".venv"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["datalabeling"]
src_paths = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "B"]
exclude = ["build", "dist", ".venv"]
src = ["src"]
